Update VITE_APP_VERSION and enhance Visit Summary dialog with new features
- Updated VITE_APP_VERSION to 0.1_20251219.
- Added a "Questionnaires & Surveys" section to the Visit Summary dialog, displaying all questionnaires with response details.
- Enhanced PDF export functionality to include complete questionnaire responses.
- Refactored Visit Summary dialog into smaller, maintainable components for improved code organization.
