+++ db/app2/.env
@@ -1,1 +1,1 @@
-VITE_APP_VERSION=0.1_20251218
+VITE_APP_VERSION=0.1_20251219
